+++ src/Timestamps.jl
@@ -1,4 +1,4 @@
-using Dates
+using Dates, Requires
 
 module Timestamps
 
@@ -6,7 +6,7 @@ using Dates, Requires
 
 import Base: show, convert, sum, prod, mean, var, std, maximum, minimum
 
-export Timestamp
+export Timestamp, TimestampArray
 
 include("timestamp.jl")
 include("conversion.jl")