+++ src/conversion.jl
@@ -24,13 +24,17 @@
 
     """
 
-function convert(::Type{Array{Timestamps.Timestamp}}, A::TimeSeries.TimeArray)
-    #stamps = Array(Timestamp, length(A))
-    stamps = Timestamp[]
+#function convert{T}(::Type{Array{Timestamp{T}}}, A::TimeSeries.TimeArray{T,1})
+function convert{T}(::Type{TimestampArray}, A::TimeSeries.TimeArray{T,1})
+    stamps = Array(Timestamp, length(A))
     for i in 1:length(A)
-        push!(stamps, Timestamp(A.timestamp[i], A.values[i]))
-        # stamps[i] = Timestamp(A.timestamp[i], A.values[i])
+        stamps[i] = Timestamp(A.timestamp[i], A.values[i])
     end
     stamps
-    end
 end
+
+# convert{T}(at::Array{Timestamp{T},1}, ta::TimeSeries.TimeArray{T,1}) = convert(at, ta) 
+#Timestamp{T,N}(ta::TimeSeries.TimeArray{T,N}) = convert{T}(Array{Timestamp}, A::TimeSeries.TimeArray{T})
+
+end # of begin statement
+