+++ src/timestamp.jl
@@ -5,6 +5,8 @@ immutable Timestamp{T} <: AbstractTimestamp
     value::T
 end
 
+typealias TimestampArray{T} Array{Timestamp{T},1}
+
 function Base.show{T}(io::IO, ts::Timestamp{T})
     print(io, ts.timestamp, " | ", ts.value)
 end