+++ NEWS.md
@@ -1,3 +1,9 @@
+### 0.0.4 (not fully implemented)
+
+* typealias TimestampArray{T} Array{Timestamp{T},1} added
+* fixes TimeArray conversion methods
+* merge and sort methods added
+
 ### 0.0.3
 
 * limit support to Julia 0.3 only